now outputs to 'Eprices.csv' and 'Aprices.csv'!
changed binomial() to return double;
fixed a thing which was returning 'h' as 0, making 'discount' -0;

+++ binomial.h
@@ -1,6 +1,7 @@
 #ifndef BINOMIAL_H_INCLUDED
 #define BINOMIAL_H_INCLUDED
 
+
 /** this file gives a function to calculate the binomial coefficents */
 
 
@@ -10,10 +11,10 @@
     @param k the number of things to choose
     @return the number of ways of choosing k things from a set of n things
 */
-long long binomial(int n, int k){
-    long long ans =1;
+double binomial(int n, int k){
+    double ans =1;
     if (k>n-k) k=n-k;          //  optimisation, using the symmetry of the
-    for (int i =0; i<k; i++){ //  binomial coefficients: C(n,k)=C(n,n-k)
+    for (double i =0; i<k; i++){ //  binomial coefficients: C(n,k)=C(n,n-k)
          ans = ans*(n-i);
          ans = ans/(i+1);
     }
